Simplify the Search code
